Conflicts:
	liquibase/Bahmni/BahmniConfig.xml
	openelis/src/us/mn/state/health/lims/common/servlet/startup/StartStopListener.java
	openelis/src/us/mn/state/health/lims/result/action/ResultsLogbookUpdateAction.java
	openelis/src/us/mn/state/health/lims/resultvalidation/action/ResultValidationSaveAction.java
	openelis/src/us/mn/state/health/lims/testanalyte/daoimpl/TestAnalyteTestResultDAOImpl.java

+++ openelis/src/us/mn/state/health/lims/siteinformation/daoimpl/SiteInformationDAOImpl.java
@@ -210,7 +210,7 @@ public class SiteInformationDAOImpl extends BaseDAOImpl implements SiteInformati
 		try {
 			Query query = HibernateUtil.getSession().createQuery(sql);
 			query.setString("domainName", domainName);
-			Integer count = (Integer)query.uniqueResult();
+			Integer count = ((Long)query.uniqueResult()).intValue();
 			closeSession();
 			return count;
 		} catch (HibernateException e) {
